Add React Navigation stack config
