+++ .eslintrc.json
@@ -26,6 +26,7 @@
   ],
   "rules": {
     "max-len": "off",
+    "react/style-prop-object": "off",
     "react/jsx-one-expression-per-line": "off",
     "react/jsx-props-no-spreading": "off",
     "react/prop-types": "off",
