+++ package.json
@@ -11,6 +11,9 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@react-native-masked-view/masked-view": "^0.2.2",
+    "@react-navigation/native": "^5.9.3",
+    "@react-navigation/stack": "^5.14.3",
     "expo": "~40.0.0",
     "expo-splash-screen": "^0.9.0",
     "expo-status-bar": "~1.0.3",
@@ -18,9 +21,10 @@
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-native": "~0.63.4",
-    "react-native-gesture-handler": "~1.8.0",
-    "react-native-reanimated": "~1.13.0",
-    "react-native-screens": "~2.15.0",
+    "react-native-gesture-handler": "^1.8.0",
+    "react-native-reanimated": "^1.13.2",
+    "react-native-safe-area-context": "^3.1.9",
+    "react-native-screens": "^2.15.2",
     "react-native-splash-screen": "^3.2.0",
     "react-native-unimodules": "~0.12.0",
     "react-native-web": "~0.13.12"
